+++ mixer/blender_data/specifics.py
@@ -227,12 +227,12 @@ def bpy_data_ctor_images(collection_name: str, proxy: DatablockProxy, context: C
     filepath = proxy.data("filepath")
 
     resolved_filepath = proxy._filepath_raw
-    if not proxy._is_in_workspace:
+    if not proxy._is_in_shared_folder:
         resolved_filepath = get_cache_file_path(proxy._filepath_raw)
     else:
-        resolved_filepath = proxy.resolve_workspace_file(proxy._filepath_raw, context)
+        resolved_filepath = proxy.resolve_shared_folder_file(proxy._filepath_raw, context)
         if resolved_filepath is None:
-            logger.warning(f'"{proxy._filepath_raw}" not in workspace')
+            logger.warning(f'"{proxy._filepath_raw}" not in shared_folder')
             return None
 
     packed_files = proxy.data("packed_files")